% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.networks.R
\name{assert_combined_network}
\alias{assert_combined_network}
\title{Confirm that this is a combined network.}
\usage{
assert_combined_network(nw, combiners, term = TRUE, call = rlang::caller_env())
}
\arguments{
\item{nw}{a \code{\link[network:network]{network}}}

\item{combiners}{a character vector listing valid combiners supported by the function}

\item{term}{whether this is a term that should use \code{\link[ergm:ergm-errors]{ergm_Init_stop()}} or some other function, which should use \code{\link[rlang:abort]{rlang::abort()}}}

\item{call}{for \code{term_trace == FALSE} where to report the error message from}
}
\description{
Confirm that this is a combined network.
}
\keyword{internal}
