% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer.R
\name{assert_LHS_Layer}
\alias{assert_LHS_Layer}
\title{Confirm that this is a multilayer network.}
\usage{
assert_LHS_Layer(nw, errfn = ergm_Init_stop)
}
\arguments{
\item{nw}{a \code{\link[network:network]{network}}}

\item{errfn}{function to call with the error message}
}
\value{
\code{TRUE} if a valid multilayer network, results of calling \code{errfn} otherwise.
}
\description{
Confirm that this is a multilayer network.
}
