% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{b2degrangeL-ergmTerm}
\alias{b2degrangeL-ergmTerm}
\title{Degree range for the second mode in a bipartite network}
\usage{
# binary: b2degrangeL(from, to=+Inf, by=NULL, homophily=FALSE, levels=NULL)
}
\arguments{
\item{from, to}{vectors of distinct
integers (or \code{+Inf} , for \code{to} (its default)). If one of the vectors has
length 1, it is recycled to the length of the other. Otherwise, they
must have the same length.}

\item{by}{an optional character string giving the name of an attribute in the
network's vertex attribute list.}

\item{homophily}{Only used if \code{by} is specified. If set to \code{TRUE} ,
then degrees are calculated using the subnetwork consisting of only
edges whose endpoints have the same value of the \code{by} attribute.
Otherwise (the default), then separate degree range
statistics are calculated for nodes having each separate
value of the attribute.}

\item{levels}{TODO}
}
\description{
Degree range for the second mode in a bipartite (a.k.a. two-mode) network
}
\details{
This term adds one
network statistic to the model for each element of \code{from} (or \code{to} ); the \eqn{i} th
such statistic equals the number of nodes of the second mode
("events") in the network of degree
greater than or equal to \code{from[i]} but strictly less than \code{to[i]} , i.e. with
edge count in semiopen interval \verb{[from,to)} .

This term can only be used with bipartite networks; for directed networks
see \code{idegrange} and \code{odegrange} . For undirected networks,
see \code{degrange} , and see \code{b1degrange} for degrees of the first mode ("actors").
}
\concept{bipartite}
\concept{undirected}
