% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{Bernoulli-ergmReference}
\alias{Bernoulli-ergmReference}
\title{Bernoulli-reference ERGM}
\usage{
# Bernoulli
}
\description{
Bernoulli-reference ERGM
}
\details{
Specifies each
dyad's baseline distribution to be Bernoulli with probability of
the tie being \eqn{0.5} . This is the only reference measure used
in binary mode.
}
