% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{blacklist_block-ergmConstraint}
\alias{blacklist_block-ergmConstraint}
\alias{InitErgmConstraint.blacklist_block}
\title{Blacklist blocks of dyads from toggling in a way that
propagates through combined networks}
\usage{
# blacklist_block(block_vattr = ".ubid",
#                 blacklist_nattr = ".block_blacklist")
}
\arguments{
\item{block_vattr}{a character string giving the name of vertex
attribute containing globally unique vertex block membership IDs.}

\item{blacklist_nattr}{a character string giving the name of a
network attribute containing the list of pairs of vertex IDs that
are to be blacklisted.}
}
\description{
This constraint is primarily for internal use,
alongside the block blacklist API implemented in
\code{block_blacklist.R}, which assigns and interprets vertex and
network attributes allowing certain blocks of dyads to be
blacklisted in a way that will survive even network manipulations
such as subgraph extraction and combining into block-diagonal
using \code{\link[=combine_networks]{combine_networks()}}.
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
\keyword{internal}
