% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.blockdiag.R
\name{blockdiagTNT-ergmProposal}
\alias{blockdiagTNT-ergmProposal}
\alias{InitErgmProposal.blockdiagTNT}
\title{A Metropolis--Hastings proposal for diagonal block constraints}
\description{
Typically used for \eqn{constraints=
  ~blockdiag}. Similar to InitErgmProposal.blockdiag, except that
it selects ties and non-ties for proposed toggles (in the block
by construction) with equal probability.  Like the unconstrained
TNT proposal, this is useful for improving performance in sparse
networks.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="blockdiagTNT"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="blockdiagTNT"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="blockdiagTNT"))}}
}
\keyword{internal}
