% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.networks.R, R/print.combined_networks.R
\name{combine_networks}
\alias{combine_networks}
\alias{combined_networks}
\alias{uncombine_network}
\alias{print.combined_networks}
\alias{summary.combined_networks}
\alias{print.summary.combined_networks}
\title{A single block-diagonal network created by combining multiple networks}
\usage{
combine_networks(
  nwl,
  ignore.nattr = c("mnext"),
  ignore.vattr = c(),
  ignore.eattr = c(),
  keep.unshared.attr = FALSE
)

uncombine_network(nw, populate = TRUE)

\method{print}{combined_networks}(x, ...)

\method{summary}{combined_networks}(object, ...)

\method{print}{summary.combined_networks}(x, ...)
}
\arguments{
\item{nwl}{a list of \code{\link[network:network]{network::network}}s to be combined; they
must have similar fundamental properties: directedness and
bipartedness, though their sizes (and the size of each bipartite
group) can vary.}

\item{ignore.nattr, ignore.vattr, ignore.eattr}{network, vertex, and
edge attributes not to be processed as described below.}

\item{keep.unshared.attr}{whether to keep those network, vertex,
and edge attributes not shared by all networks in the list; if
\code{TRUE}, positions corresponding to networks lacking the
attribute are replaced with \code{NA}, \code{NULL}, or some other
placeholder; incompatible with \code{detect.edgecov==TRUE}.}

\item{nw}{a \code{\link[network:network]{network::network}} created by \code{\link[=combine_networks]{combine_networks()}}.}

\item{populate}{logical; whether to map the edges from the combined
network to the subnetworks, or whether to return them empty.}

\item{x, object}{a combined network.}

\item{...}{additional arguments to methods.}
}
\value{
an object of class \code{combined_networks} inheriting from
\code{\link[network:network]{network::network}} with a block-diagonal structure (or its
bipartite equivalent) comprising the networks passed in \code{nwl}. In
particular,
\itemize{
\item the returned network's size is the sum of the input networks';
\item its basic properties (directedness and bipartednes) are the same;
\item the input networks' sociomatrices (both edge presence and edge
attributes) are the blocks in the sociomatrix of the returned
network;
\item vertex attributes are concatenated;
\item edge attributes are assigned to their respective edges in
the returned network;
\item network attributes are stored in a list network attribute \code{".snattr"}.
\item a new network attribute, ".snid" identifies the original
networks; if combined networks are combined again, ".snid"
becomes a space-separated string of IDs from outside in.
\item a  new network  attribute, ".snl",  containing networks  in \code{nwl}
(preserving their names, if any) but with all edges removed.
}
\enumerate{
\item If \code{nwl} is a named list, names from the list are used.
\item If not 1, but the network has an attribute \code{title}, it is used.
\item Otherwise, a numerical index is used.
}
}
\description{
Given a list of compatible networks, \code{\link[=combine_networks]{combine_networks()}} returns a single
block-diagonal network, preserving attributes that can be
preserved.

\code{\link[=uncombine_network]{uncombine_network()}} undoes \code{\link[=combine_networks]{combine_networks()}}, returning a list
of networks, optionally mapping the edges of the combined network
onto them.
}
\section{Functions}{
\itemize{
\item \code{print(combined_networks)}: A wrapper around
\code{\link[network:network]{network::print.network()}} to print constituent network
information and omit some internal variables.

\item \code{summary(combined_networks)}: A wrapper around
\code{\link[network:network]{network::summary.network()}} to print constituent network
information and omit some internal variables.

\item \code{print(summary.combined_networks)}: A wrapper around
\code{\link[network:network]{network::print.summary.network()}} to print constituent network
information and omit some internal variables.

}}
\examples{

data(samplk)

o1 <- combine_networks(list(samplk1, samplk2, samplk3))
image(as.matrix(o1))
head(get.vertex.attribute(o1, ".snid"))
o2 <- combine_networks(list(o1, o1))
image(as.matrix(o2))
head(get.vertex.attribute(o2, ".snid", unlist=FALSE))

data(florentine)
f1 <- combine_networks(list(business=flobusiness, marriage=flomarriage))
image(as.matrix(f1))
head(get.vertex.attribute(f1, ".snid"))
head(names(get.network.attribute(f1, ".snl")))

data(samplk)

o1 <- combine_networks(list(samplk1, samplk2, samplk3))
image(as.matrix(o1))

ol <- uncombine_network(o1)

}
\seealso{
\code{\link[=split.network]{split.network()}}
}
