% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{degreeL-multilayer-ergmTerm}
\alias{degreeL-multilayer-ergmTerm}
\alias{InitErgmTerm.degreeL}
\title{Layer combinaton degree}
\usage{
# binary: degreeL(d, by=NULL, Ls, dir)
}
\arguments{
\item{d}{a vector of distinct integers}

\item{by}{if specified, then separate degree
statistics are calculated for nodes having each separate
value of the attribute.}

\item{dir}{vector specifying direction}

\item{Ls}{either a Layer Logic specification formula (c.f. Layer Logic section in the \code{\link[=Layer]{Layer()}} documentation) or a list thereof (constructed by \code{\link[=list]{list()}} or \code{\link[=c]{c()}}). If given, degree of a node}
}
\description{
This term adds one network statistic to
the model for each element in \code{d} ; the \eqn{i} th such statistic equals
the number of nodes in the network whose total degree across all
layers listed in \code{Ls} combined (with direction being
corresponding element of vector \code{dir} , \code{"in"} or \code{"out"} ), is \code{d[i]} .

This term can only be used with directed networks; for undirected networks
see \code{degree} .
}
\concept{categorical nodal attribute}
\concept{directed}
\concept{frequently-used}
