% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.dgw_sp.R
\name{despL-ergmTerm}
\alias{despL-ergmTerm}
\alias{InitErgmTerm.despL}
\title{Directed edgewise shared partners}
\usage{
# binary: despL(d, type="OTP", L.base=NULL, Ls.path=NULL, L.in_order=FALSE)
}
\arguments{
\item{d}{a vector of distinct integers}

\item{type}{select the type of shared partner to be counted (see \code{ddsp} for type codes). By default, outgoing two-paths are employed.}

\item{L.base}{the layer specification for the base}

\item{Ls.path, L.in_order}{a vector of one or two formulas
\code{Ls.path} provides the layer specifications for the ties of the
2-path or the shared partnership. (If only one formula is given the
layers are assumed to be the same.) If \code{L.in_order==TRUE} , the
first tie of the two-path must be the first element of \code{Ls.path}
and the second must be the second; otherwise, any ordering counts,
provided there is exactly one of each. (For types \code{"OSP"} and \code{"ISP"} , the first tie is considered to be the one incident on the tail of the base tie.)}
}
\description{
This term adds one network statistic to the model for each element in \code{d} where the \eqn{i} th such statistic equals the number of edges in the network with exactly \code{d[i]} shared partners. This term can only be used with directed networks. Multiple shared partner definitions are possible.
}
\concept{directed}
\concept{layer-aware}
\concept{undirected}
