% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.dgw_sp.R
\name{dgwdspL-ergmTerm}
\alias{dgwdspL-ergmTerm}
\alias{InitErgmTerm.dgwdspL}
\title{Geometrically weighted dyadwise shared partner distribution}
\usage{
# binary: dgwdspL(decay, fixed=FALSE, cutoff=30, type="OTP",
#                 Ls.path=NULL, L.in_order=FALSE)
}
\arguments{
\item{decay}{non-negative model parameter (this parameter was called \code{alpha} prior to \verb{ergm 3.7} ).}

\item{fixed}{logical to specify if \code{decay} should be fixed (if \code{fixed=TRUE} ), or it may be used instead as the starting value for the estimation of \code{decay} in a curved exponential family model (when \code{fixed=FALSE} , the default) (see Hunter and Handcock, 2006).}

\item{cutoff}{optional argument to set the number of underlying DSP terms to use in computing the statistics to reduce the computational burden. Its default value can also be controlled by the \code{gw.cutoff} term option control parameter. (See \code{\link{control.ergm}} .)}

\item{type}{select the type of shared partner to be counted (see \code{ddsp} for type codes). By default, outgoing two-paths are employed.}

\item{Ls.path, L.in_order}{a vector of one or two formulas
\code{Ls.path} provides the layer specifications for the ties of the
2-path or the shared partnership. (If only one formula is given the
layers are assumed to be the same.) If \code{L.in_order==TRUE} , the
first tie of the two-path must be the first element of \code{Ls.path}
and the second must be the second; otherwise, any ordering counts,
provided there is exactly one of each. (For types \code{"OSP"} and \code{"ISP"} , the first tie is considered to be the one incident on the tail of the base tie.)}
}
\description{
Geometrically weighted dyadwise shared partner distribution
}
\details{
This term adds one network statistic to the model equal to the geometrically weighted dyadwise shared partner distribution with decay parameter. Note that the GWDSP statistic is equal to the sum of GWNSP plus GWESP. For a directed network, multiple shared partner definitions are possible; the \code{type} argument may be used to select the type of shared partner to employ (see \code{ddsp} for definitions). By default, outgoing two-paths are employed.

See term \code{ddsp} for layer specification information.
}
\concept{directed}
\concept{layer-aware}
\concept{undirected}
