% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{DiscUnif-ergmReference}
\alias{DiscUnif-ergmReference}
\title{Discrete-Uniform-reference ERGM}
\usage{
# DiscUnif(a,b)
}
\description{
Discrete-Uniform-reference ERGM
}
\details{
Specifies each dyad's baseline distribution to be discrete uniform
between \code{a} and \code{b} (both inclusive): \eqn{h(y)=1} , with
the support being
\code{a} , \code{a} +1, , \code{b} -1, \code{b} . At this time, both \code{a} and
\code{b} must be finite.
}
