% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{entrainmentL-ergmTerm}
\alias{entrainmentL-ergmTerm}
\alias{InitErgmTerm.entrainmentL}
\title{Entrainment between pairs of lairs}
\usage{
# binary: entrainmentL(Ls=~.)
}
\arguments{
\item{Ls}{a list (constructed by \code{\link[=list]{list()}} or \code{\link[=c]{c()}} of at least two Layer Logic specifications (c.f. Layer Logic section in the \code{\link[=Layer]{Layer()}} documentation).}
}
\description{
Models marginal dependence of layers within each dyad
by counting edges they have in common.

The term adds one statistic to the model, equalling the sum over
all distinct pairs of specified layers of their common edges.
}
\details{
This is similar to \code{L(~edges, c(~A & B, ~A & C, ~B & C))}
but more efficient for multiple layers.
}
\seealso{
\code{\link[ergm:ergmTerm]{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "entrainmentL", "subsection")}
}
\concept{directed}
\concept{layer-aware}
\concept{undirected}
