% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.networks.R
\name{ergm.multi_combiner}
\alias{ergm.multi_combiner}
\title{Dynamic registry of functions that combine networks.}
\usage{
ergm.multi_combiner(combiner, description = NULL)
}
\arguments{
\item{combiner}{the combiner name (typically \code{nw \%n\% ".combiner"})}

\item{description}{a named character vector or list with elements
\code{constructor}, \code{element}, \code{elements}, \code{construct}, \code{id}, and
\code{name}, containing, respectively, the top-level
(i.e. user-visible) function that combines on it, what is being
combined (singular and plural), and what columns, if any, should
be used to identify its index and name in the attributes
table. \code{elements} and \code{construct} will be extrapolated from the
others if not given.}
}
\description{
This is used primarily by developers to provide informative error
messages.
}
\keyword{internal}
