% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergmlhs.R
\name{ergmlhs_remove_blockdiag}
\alias{ergmlhs_remove_blockdiag}
\title{Remove a \code{\link[ergm:blockdiag-ergmConstraint-140bec05]{blockdiag}} constraint for a
specific vertex attribute from an \code{\link[ergm:ergmlhs]{\%ergmlhs\%}} formula if it's
absent in the network.}
\usage{
ergmlhs_remove_blockdiag(nw, vattr)
}
\arguments{
\item{nw}{a \code{\link[network:network]{network}}}

\item{vattr}{a vertex attribute name}
}
\value{
a \code{\link[network:network]{network}} with an updated \verb{\%ergmlhs\% "constraints"}.
}
\description{
Remove a \code{\link[ergm:blockdiag-ergmConstraint-140bec05]{blockdiag}} constraint for a
specific vertex attribute from an \code{\link[ergm:ergmlhs]{\%ergmlhs\%}} formula if it's
absent in the network.
}
\keyword{internal}
