% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.multi-terms-index.R
\docType{package}
\name{ergmTerm}
\alias{ergmTerm}
\alias{ergm-terms}
\alias{ergm.terms}
\alias{terms-ergm}
\alias{terms.ergm}
\alias{InitErgmTerm}
\alias{InitErgmWtTerm}
\title{Terms used in Exponential Family Random Graph Models}
\description{
The function \code{\link{ergm}} is used to fit exponential random graph
models, in which the probability of a given network, \eqn{y}, on a set of
nodes is \deqn{h(y) \exp\{\eta(\theta) \cdot }{h(y) exp{eta(theta).g(y)} /
c(theta),}\deqn{ g(y)\}/c(\theta)}{h(y) exp{eta(theta).g(y)} / c(theta),}
where \eqn{h(y)} is the reference measure (for valued network models),
\eqn{g(y)} is a vector of network statistics for \eqn{y}, \eqn{\eta(\theta)}
is a natural parameter vector of the same length (with
\eqn{\eta(\theta)\equiv\theta}{eta(theta)=theta} for most terms),
\eqn{\cdot}{.} is the dot product, and \eqn{c(\theta)} is the normalizing
constant for the distribution.

The network statistics \eqn{g(y)} are entered as terms in the function call
to \code{\link{ergm}}.  This page describes the possible terms (and hence
network statistics) included in \code{\link[=ergm-package]{ergm}} package.

A cross-referenced HTML version of the term documentation is available via
\code{vignette('ergm-term-crossRef')} and terms can also be searched via
\code{\link{search.ergmTerms}}.
}
\section{Multilayer networks}{

In order to fit a model for multilayer
networks, first use \code{\link{Layer}} construct an LHS network that
\code{\link{ergm}} will understand as multilayered.

Used on the formula directly, most, but not all, terms included in this
package will sum their statistics over the layers.

Some terms are \emph{layer-aware}. These terms have explicit (usually
optional) layer specification arguments. For other terms, an operator
\code{L(formula, Ls=~.)} can be used to specify to which layer(s) they
apply. Layer specification documentation follows.
\subsection{Layer Logic}{

Each formula's right-hand side describes an observed layer \emph{or} some
"logical" layer, whose ties are a function of corresponding ties in
observed layers.

The observed layers can be referenced either by name or by number (i.e.,
order in which they were passed to \code{\link{Layer}}). When referencing by
number, enclose the number in quotation marks (e.g., "1") or
backticks (e.g., \dQuote{\code{1}}).

\link[base:Arithmetic]{Arithmetical}, \link[base:Comparison]{relational},
and \link[base:Logic]{logical} operators can be used to combine them. All
listed operators are implemented, as well as functions \code{\link{abs}},
\code{\link{round}}, and \code{\link{sign}}. Standard
\link[base:Syntax]{operator precedence} applies, so use of parentheses is
recommended to ensure the logical expression is what it looks like.

For example, if LHS is \code{Layer(A=nwA, B=nwB)}, both \code{~`2`} and
\code{~B} refer to \code{nwB}, while \code{A&!B} refers to a
\dQuote{logical} layer that has ties that are in \code{nwA} but not in
\code{nwB}.

Transpose function \code{\link{t}} applied to a directed layer will reverse
the direction of all relations (transposing the sociomatrix). Unlike the
others, it can only be used on an observed layer directly. For example,
\code{~t(`1`)&t(`2`)} is valid but \code{~t(`1`&`2`)} is not.

At this time, logical expressions that produce complete graphs from empty
graph inputs (e.g., \code{A==B} or \code{!A}) are not supported.
}

\subsection{Summing layers}{

Some terms combine multiple layers. I that case al ist of formulas may be
passed. For example, \code{Layer(nw1,nw2)~L(~edges, c(~`1`,~(`2`&!`1`)))}
produces the number of edges in layer 1 plus the number of edges in layer 2
but not in layer 1.

For these formulas, one can specify the layer's weight on its left-handside.
For example, \code{Layer(nw1,nw2)~L(~edges, c(3~`1`,-1~(`2`&!`1`)))} will
produce three times the number of edges in layer 1, minus the number of
edges in layer 2 but not in layer 1.
}
}

\section{Nodal attribute levels}{

Terms taking a categorical nodal covariate also take \code{levels}
argument. This can be used to control the set and the ordering of
attribute levels.
}

\section{Terms included in the \code{\link[=ergm-package]{ergm.multi}} package}{

As noted above, a cross-referenced HTML version of the term documentation is
also available via \code{vignette('ergm-term-crossRef')} and terms
can also be searched via \code{\link{search.ergmTerms}}.
\subsection{Term index}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexLatex(ergm:::.buildTermsDataframe("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexText(ergm:::.buildTermsDataframe("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexHtml(ergm:::.buildTermsDataframe("ergmTerm"))}}
}

\subsection{Frequently-used terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
}

\subsection{Operator terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
}

\subsection{All terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm"))}}
}

\subsection{Terms by concept}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocLatex(ergm:::.termToc("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocText(ergm:::.termToc("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocHtml(ergm:::.termToc("ergmTerm"))}}
}
}

\examples{
\dontrun{
ergm(flomarriage ~ kstar(1:2) + absdiff("wealth") + triangle)

ergm(molecule ~ edges + kstar(2:3) + triangle
                      + nodematch("atomic type",diff=TRUE)
                      + triangle + absdiff("atomic type"))
}
}
\references{
\itemize{
\item Butts, CT.  (2008).  "A Relational Event Framework for Social
Action." \emph{Sociological Methodology,} 38(1).
\item Davis, J.A. and Leinhardt, S.  (1972).  The Structure of Positive
Interpersonal Relations in Small Groups.  In J. Berger (Ed.),
\emph{Sociological Theories in Progress, Volume 2}, 218--251.  Boston:
Houghton Mifflin.
\item Holland, P. W. and S. Leinhardt (1981). An exponential family of
probability distributions for directed graphs.  \emph{Journal of the
American Statistical Association}, 76: 33--50.
\item Hunter, D. R. and M. S. Handcock (2006). Inference in curved
exponential family models for networks. \emph{Journal of Computational and
Graphical Statistics}, 15: 565--583.
\item Hunter, D. R. (2007). Curved exponential family models for social
networks. \emph{Social Networks}, 29: 216--230.
\item Krackhardt, D. and Handcock, M. S. (2007).  Heider versus Simmel:
Emergent Features in Dynamic Structures. \emph{Lecture Notes in Computer
Science}, 4503, 14--27.
\item Krivitsky P. N. (2012). Exponential-Family Random Graph Models for
Valued Networks. \emph{Electronic Journal of Statistics}, 2012, 6,
1100-1128. \doi{10.1214/12-EJS696}
\item Robins, G; Pattison, P; and Wang, P.  (2009).  "Closure,
Connectivity, and Degree Distributions: Exponential Random Graph (p*) Models
for Directed Social Networks." \emph{Social Networks,} 31:105-117.
\item Snijders T. A. B., G. G. van de Bunt, and C. E. G. Steglich.
Introduction to Stochastic Actor-Based Models for Network Dynamics.
\emph{Social Networks}, 2010, 32(1), 44-60. \doi{10.1016/j.socnet.2009.02.004}
\item Morris M, Handcock MS, and Hunter DR. Specification of
Exponential-Family Random Graph Models: Terms and Computational Aspects.
\emph{Journal of Statistical Software}, 2008, 24(4), 1-24.
\url{https://www.jstatsoft.org/v24/i04}
\item Snijders, T. A. B., P. E. Pattison, G. L. Robins, and M. S. Handcock
(2006). New specifications for exponential random graph models,
\emph{Sociological Methodology}, 36(1): 99-153.
Sample Space Constraints for Exponential-Family Random Graph Models
}
}
\seealso{
\code{\link[=ergm-package]{ergm}} package, \code{\link{search.ergmTerms}}, \code{\link{ergm}}, \code{\link{network}}, \code{\link{\%v\%}}, \code{\link{\%n\%}}
}
\keyword{models}
