% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.networks.R
\name{get_combining_attr}
\alias{get_combining_attr}
\title{In a combined network, obtain the top-level attribute vector on which it was combined.}
\usage{
get_combining_attr(x, attrname, missing = c("NA", "NULL"))
}
\arguments{
\item{x}{a \code{\link[=combine_networks]{combine_networks()}} network}

\item{attrname}{a string}

\item{missing}{what to return if the attribute is not found}
}
\description{
In a combined network, obtain the top-level attribute vector on which it was combined.
}
\keyword{internal}
