% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Goeyvaerts}
\alias{Goeyvaerts}
\title{A sample of within-household contact networks in Flanders and Brussels}
\format{
An object of class \code{list} of length 318.
}
\source{
The data were collected and by Goeyvaerts et al. and
curated by Pietro Coletti.
}
\usage{
data(Goeyvaerts)
}
\description{
This is a list of 318 \code{\link{network}} objects derived from contact
diary data collected by by Goeyvaerts et al. (2018). The study
recruited households in Flanders and Brussels-Capital region with
at least one child 12 or under. The networks are symmetrized.
}
\section{Nonstandard Network Attributes}{
 \describe{

\item{\code{included}}{(logical) whether the network was included in
Goeyvaerts's analysis. (Two were excluded.)}

\item{\code{weekday}}{(logical) whether the contact diary on which the
network is based was collected on a weekday, as opposed to
weekend.}

}
}

\section{Nonstandard Vertex Attributes}{
 \describe{

\item{\code{age}}{(numeric) the household member's age.}

\item{\code{gender}}{(character) the household member's gender (\code{"F"}/\code{"M"}).}

\item{\code{role}}{(character) the household member's inferred role (\code{"Father"}/\code{"Mother"}/\code{"Child"}/\code{"Grandmother"}).}

}
}

\section{Licenses and Citation}{
 When publishing results obtained using this
data set, the original authors (Goeyvaerts et al., 2018)
should be cited, along with this \R package.
}

\references{
Nele Goeyvaerts, Eva Santermans, Gail Potter, Andrea Torneri, Kim
V. Kerckhove, Lander Willem, Marc Aerts, Philippe Beutels & Niel
Hens (2018) Household Members Do Not Contact Each Other at Random:
Implications for Infectious Disease Modelling. \emph{Proceedings of the
Royal Society B: Biological Sciences}, 285(1893):
20182201. \doi{10.1098/rspb.2018.2201}
}
\seealso{
\code{vignette("Goeyvaerts_reproduction")} for a vignette reproducing the Goeyvaerts analysis and performing diagnostics
}
\keyword{datasets}
