% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitErgmTerm.multilayer.R
\name{gwdegreeL-ergmTerm}
\alias{gwdegreeL-ergmTerm}
\alias{InitErgmTerm.gwdegreeL}
\title{Geometrically weighted degree distribution}
\usage{
# binary: gwdegreeL(decay, fixed=FALSE, attrname=NULL, cutoff=30, levels=NULL)

# binary: gwdegreeL(decay, fixed=FALSE, attrname=NULL, cutoff=30, levels=NULL, Ls, dir)
}
\arguments{
\item{decay}{non negative model parameter (this parameter was called \code{alpha} prior to \verb{ergm 3.7} )}

\item{fixed}{if \code{TRUE}, use the value specified by \code{decay}. Otherwise (the default) use the value as the starting value for the estimation of \code{decay}
in a curved exponential family model (see Hunter and Handcock, 2006).}

\item{attrname}{if specified then separate degree
statistics are calculated for nodes having each separate
value of the attribute.}

\item{cutoff}{optional argument that is only relevant if \code{fixed=FALSE} . In that case it only uses this
number of terms in computing the statistics to reduce the computational
burden. Its default value can also be controlled by the \code{gw.cutoff} term option control parameter. (See \code{\link{control.ergm}} .)}

\item{levels}{TODO}

\item{Ls}{layers specification}

\item{dir}{vector of directions}
}
\description{
This term adds one network statistic to the model equal to the weighted
degree distribution with decay controlled by the \code{decay} parameter.

This term can only be used with undirected networks.

This term adds one network statistic to the model
equal to the weighted layer-cumulative degree distribution with decay parameter. This
term can only be used with directed networks.

In this term, the layer-cumulative degree frequency equals
the number of nodes in the network whose total degree across all
layers listed in \code{Ls} combined (with direction being
corresponding element of vector \code{dir} , \code{"in"} or \code{"out"} ), is \code{d[i]} .
}
\concept{curved}
\concept{directed}
\concept{frequently-used}
\concept{undirected}
