% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{hammingL-ergmTerm}
\alias{hammingL-ergmTerm}
\alias{InitErgmTerm.hammingL}
\title{Hamming distance between pairs of lairs}
\usage{
# binary: hammingL(Ls=~.)
}
\arguments{
\item{Ls}{a list (constructed by \code{\link[=list]{list()}} or \code{\link[=c]{c()}} of at least two Layer Logic specifications (c.f. Layer Logic section in the \code{\link[=Layer]{Layer()}} documentation).}
}
\description{
Models marginal dependence of layers within each dyad
by counting their hamming distances.

The term adds one statistic to the model, equalling the sum over
all distinct pairs of specified layers of their hamming distances.
}
\details{
A positive coefficient induces negative dependence and a negative
one induces positive dependence.
}
\seealso{
\code{\link[ergm:ergmTerm]{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "hammingL", "subsection")}
}
\concept{directed}
\concept{layer-aware}
\concept{undirected}
