% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{L-ergmTerm}
\alias{L-ergmTerm}
\alias{InitErgmTerm.L}
\title{Evaluation on layers}
\usage{
# binary: L(formula, Ls=~.)
}
\arguments{
\item{formula}{a one-sided \code{\link[ergm:ergm]{ergm()}}-style formula with the terms to be evaluated}

\item{Ls}{either a Layer Logic specification formula (c.f. Layer Logic section in the \code{\link[=Layer]{Layer()}} documentation) or a list thereof (constructed by \code{\link[=list]{list()}} or \code{\link[=c]{c()}}), on which to evaluate \code{formula}}
}
\description{
Evaluates the terms in \code{formula} on an observed or
logical layers specified in formula \code{Ls} and sums the results
elementwise.
}
\note{
For the purposes of the terms in the \code{formula}, nonstandard
network and vertex attributes will be taken from the \emph{first}
layer's network. The subsequent networks' attributes will be
ignored by it. If \code{Ls} is specified as a \code{\link{function}}, this
function will be able to access them, however.
}
\seealso{
\code{\link[ergm:ergmTerm]{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "L", "subsection")}

\code{\link[=Layer]{Layer()}}
}
\concept{layer-aware}
\concept{operator}
