% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{L-ergmTerm}
\alias{L-ergmTerm}
\alias{InitErgmTerm.L}
\title{Evaluation on layers}
\usage{
# binary: L(formula, Ls=~.)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated}

\item{Ls}{either a Layer Logic specification formula (c.f. Layer Logic section in the \code{\link[=Layer]{Layer()}} documentation) or a list thereof (constructed by \code{\link[=list]{list()}} or \code{\link[=c]{c()}}), on which to evaluate \code{formula}}
}
\description{
Evaluates the terms in \code{formula} on an observed or
logical layers specified in formula \code{Ls} and sums the results
elementwise.
}
\concept{layer-aware}
\concept{operator}
