% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multinet.R,
%   R/InitWtErgmTerm.multinet.R
\name{N-ergmTerm}
\alias{N-ergmTerm}
\alias{InitErgmTerm.N}
\alias{InitWtErgmTerm.N}
\title{Evaluation on multiple networks}
\usage{
# binary: N(formula, lm=~1, subset=TRUE, weights=1, contrasts=NULL, offset=0, label=NULL)

# valued: N(formula, lm=~1, subset=TRUE, weights=1, contrasts=NULL, offset=0, label=NULL)
}
\arguments{
\item{offset}{A constant, a vector of length equal to the number of networks, or a matrix whose number of rows is the number of networks and whose number of columns is the number of free parameters of the ERGM. It can be specified in \code{lm} as well.}

\item{weights}{reserved for future use; attempting to change it will cause an
error: at this time, there is no way to assign sampling weights to
networks.}

\item{label}{An optional parameter which will add a label to model
parameters to help identify the term (which may have similar
predictors but, say, a different network subset) in the output.}
}
\description{
Evaluation on multiple networks
}
\details{
Evaluates the terms in \code{formula} on each of the networks joined
using \code{\link{Networks}} function, and returns either a weighted
sum or an \code{\link{lm}} -style linear model for the ERGM
coefficients. Its syntax follows that of \code{\link{lm}} closely,
with sensible defaults. The default formula ( \code{~1} ) sums the
specified network statistics. If \code{lm} refers to any
network attributes for which some networks have missing values, the
term will stop with an error. This can be avoided by pre-filtering
with \code{subset} , which controls which networks are affected by
the term.

Care should be taken to avoid multicollinearity when using this operator. When \code{lm} is given a model with intercept and a categorical predictor (including a \code{\link{logical}} one), it will use the first level (or \code{FALSE} ) as the baseline, but if the model is without intercept, it will use all levels of the first categorical predictor. This is typically what is wanted in a linear regression, but for the \code{N} operator, this can be problematic if the "intercept" effect is added by a different term. A workaround is to convert the categorical predictor to dummy variables before putting it into the \code{lm} formula.
}
\concept{operator}
