% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{network.layercount}
\alias{network.layercount}
\title{Number of layers in a combined network.}
\usage{
network.layercount(x, ...)
}
\arguments{
\item{x}{a combined network returned by \code{\link[=combine_networks]{combine_networks()}}}
}
\description{
Number of layers in a combined network.
}
\keyword{internal}
