% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{StdNormal-ergmReference}
\alias{StdNormal-ergmReference}
\title{Standard-Normal-reference ERGM}
\usage{
# StdNormal
}
\description{
Standard-Normal-reference ERGM
}
\details{
Specifies each dyad's baseline distribution to be the normal distribution
with mean 0 and variance 1.
}
