% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.networks.R
\name{subnetwork_templates}
\alias{subnetwork_templates}
\title{Obtain empty networks representing constituents of a combined network}
\usage{
subnetwork_templates(
  nw,
  split.vattr = nw \%n\% ".blockID.vattr",
  names.vattr = nw \%n\% ".blockName.vattr",
  copy.ergmlhs = c("response")
)
}
\arguments{
\item{nw, split.vattr, names.vattr}{see \code{\link[=uncombine_network]{uncombine_network()}}.}

\item{copy.ergmlhs}{a character vector of \code{\link{\%ergmlhs\%}} settings that are to be copied into the constituent networks.}
}
\value{
A list of \code{\link{network}}s.
}
\description{
This utility uncombines a \code{\link[=combine_networks]{combine_networks()}} network using subnetwork cache (which contains only empty networks). It is used primarily by initialisation functions.
}
\keyword{internal}
