% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{Unif-ergmReference}
\alias{Unif-ergmReference}
\title{Coninuous-Uniform-reference ERGM}
\usage{
# Unif(a,b)
}
\description{
Coninuous-Uniform-reference ERGM
}
\details{
Specifies each dyad's baseline distribution to be continuous uniform
between \code{a} and \code{b} (both inclusive): \eqn{h(y)=1} , with
the support being
\verb{[a, b]}. At this time, both \code{a} and
\code{b} must be finite.
}
