% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{upper_tri-ergmConstraint}
\alias{upper_tri-ergmConstraint}
\alias{InitErgmConstraint.upper_tri}
\title{Only dyads in the upper-triangle of the sociomatrix may be
toggled}
\usage{
# upper_tri(attr = NULL)
}
\arguments{
\item{attr}{quantitative attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}
}
\description{
For a directed network, only dyads \eqn{(i,j)} for
which \eqn{i < j} may be toggled. Optional argument \code{attr}
controls which subgraphs are thus restricted.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.
}
\concept{directed}
\concept{dyad-independent}
