% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.networks.R
\name{ergm.multi_combiner}
\alias{ergm.multi_combiner}
\title{Dynamic registry of functions that combine networks.}
\usage{
ergm.multi_combiner(blockID, combiner = NULL)
}
\arguments{
\item{blockID}{the vertex attribute used as the ID of the block.}

\item{combiner}{a character vector of length 1 or 2, giving the
top-level (i.e. user-visible) function that combines on it and
optionally what is being combined.}
}
\description{
This is used primarily by developers to provide informative error
messages.
}
\keyword{internal}
