% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer.R
\docType{data}
\name{ergm_LayerLogic}
\alias{ergm_LayerLogic}
\alias{ergm_LayerLogic.ergm_LayerLogic}
\alias{ergm_LayerLogic.default}
\alias{toString.ergm_LayerLogic}
\alias{ergm_LayerLogics}
\alias{ergm_LayerLogics.formula}
\alias{ergm_LayerLogics.character}
\alias{ergm_LayerLogics.numeric}
\alias{ergm_LayerLogics.logical}
\alias{ergm_LayerLogics.ergm_LayerLogic}
\alias{unlist.ergm_LayerLogics}
\alias{ergm_LayerLogics.list}
\alias{ergm_LayerLogics.function}
\alias{toString.ergm_LayerLogics}
\alias{[.ergm_LayerLogics}
\alias{to_ergm_Cdouble.ergm_LayerLogic}
\alias{ERGM_LAYER_SPEC}
\alias{ERGM_LAYERS_SPEC}
\title{Internal representation of Layer Logic}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
ergm_LayerLogic(x, ...)

\method{ergm_LayerLogic}{ergm_LayerLogic}(x, ...)

\method{ergm_LayerLogic}{default}(x, ...)

\method{toString}{ergm_LayerLogic}(x, ...)

ergm_LayerLogics(x, ...)

\method{ergm_LayerLogics}{formula}(x, lnw, ...)

\method{ergm_LayerLogics}{character}(x, lnw, ...)

\method{ergm_LayerLogics}{numeric}(x, lnw, ...)

\method{ergm_LayerLogics}{logical}(x, lnw, ...)

\method{ergm_LayerLogics}{ergm_LayerLogic}(x, ...)

\method{unlist}{ergm_LayerLogics}(x, recursive = TRUE, use.names = TRUE)

\method{ergm_LayerLogics}{list}(x, lnw, ...)

\method{ergm_LayerLogics}{`function`}(x, lnw, ...)

\method{toString}{ergm_LayerLogics}(x, ...)

\method{[}{ergm_LayerLogics}(x, i)

\method{to_ergm_Cdouble}{ergm_LayerLogic}(x, ...)

ERGM_LAYER_SPEC

ERGM_LAYERS_SPEC
}
\arguments{
\item{x}{An \code{ergm_LayerLogic} object.}

\item{...}{Additional arguments, currently unused.}

\item{lnw}{a \code{\link{combined_networks}} with layer metadata at the top level.}

\item{i}{logical or numeric indices.}
}
\value{
A structure with class
\code{"ergm_LayerLogic"}, comprising the input
\code{formula} and the following attributes: \describe{

\item{\code{namemap}}{a numeric vector mapping from layer names to layer IDs}

\item{\code{lnw}}{the provided network}

\item{\code{repr}}{a string for pretty-printing the expression}

\item{\code{aux}}{a formula with auxiliary specifications involved}

\item{\code{dep}}{indices of observed layers on which the specified layer logic depends}

\item{\code{C}}{an integer vector encoding the layer logic for C}

}

or a flat list of them with all of the above attributes except \code{C}.

The list version is a list of layer logics with the same attributes summarising them.
}
\description{
\code{ergm_LayerLogic()} always returns a single layer logic item;
\code{ergm_LayerLogics()} always returns a flat list of
\code{ergm_LayerLogic} items.
}
\section{Methods (by class)}{
\itemize{
\item \code{ergm_LayerLogic(ergm_LayerLogic)}: conversion to self: do nothing.

\item \code{ergm_LayerLogic(default)}: default method; calls the corresponding
\code{ergm_LayerLogics()} formula, then extracts the first element if there is only one, error otherwise.

}}
\section{Methods (by generic)}{
\itemize{
\item \code{toString(ergm_LayerLogic)}: A method to generate coefficient names
associated with the Layer Logic.

\item \code{ergm_LayerLogics(ergm_LayerLogic)}: a method for \code{\link[=ergm_LayerLogic]{ergm_LayerLogic()}} objects, simply insert it into a list.

\item \code{to_ergm_Cdouble(ergm_LayerLogic)}: A method to encode and serialize the
Layer Logic into a postfix program understood by the C code.

}}
\section{Functions}{
\itemize{
\item \code{ergm_LayerLogics()}: A generic for constructing a list of layer logics.

\item \code{ergm_LayerLogics(formula)}: a method for formulas.

\item \code{ergm_LayerLogics(character)}: method for \code{\link{character}}; tries to
convert the specifications to formulas, prepending "~" if
necessary.

\item \code{ergm_LayerLogics(numeric)}: method for \code{\link{numeric}}; uses them to index the observed layers.

\item \code{ergm_LayerLogics(logical)}: method for \code{\link{logical}}; uses them to index the observed layers.

\item \code{unlist(ergm_LayerLogics)}: a method to flatten the list.

\item \code{ergm_LayerLogics(list)}: construct from a list.

\item \code{ergm_LayerLogics(`function`)}: given a function and a \code{\link[network:network]{network}},
call the function (with arguments \code{lattr} containing a data frame
of layer attributes, \code{lnw} containing the combined network, and
\code{namemap} containing the mapping from layer names to layer
indices) to construct the formulas or an index (numeric, logical,
or character) for the subset of layers.

\item \code{toString(ergm_LayerLogics)}: A method to generate coefficient names
associated with the Layer Logics.

\item \code{[}: Indexing for lists of
\code{ergm_LayerLogic}s, updating metadata.

\item \code{ERGM_LAYER_SPEC}: variable type for format specification when one logical layer is required.

\item \code{ERGM_LAYERS_SPEC}: variable type for format specification when one or more logical layers are required.

}}
\keyword{datasets}
\keyword{internal}
