% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{L-ergmTerm}
\alias{L-ergmTerm}
\alias{InitErgmTerm.L}
\title{Evaluation on layers}
\usage{
# binary: L(formula, Ls=~.)
}
\description{
Evaluation on layers
}
\details{
Evaluates the terms in \code{formula} on an actual or logical layers
specified in formula \code{Ls} and sums the results. \code{Ls} is a
one-sided formula with terms separated by \code{+} operators, and
corresponding statistics for each layer specification is summed
up. A dot ( \code{.} ) stands for summing all layers.
}
\concept{layer-aware}
\concept{operator}
