% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{Layer}
\alias{Layer}
\title{A multilayer network representation.}
\usage{
Layer(..., .symmetric = NULL, .bipartite = NULL)
}
\arguments{
\item{...}{layer specification, in one of three formats:
\enumerate{
\item An (optionally named) list of identically-dimensioned
networks.
\item Several networks as (optionally named) arguments.
\item A single network, a character vector, and several optional
arguments. Then, the layers are values of the named edge
attributes. The optional arguments \code{.symmetric} and
\code{.bipartite} are then interpreted as described below.
}}

\item{.symmetric}{If the layer specification is via a single
network with edge attributes and the network is directed, an
optional logical vector to specify which of the layers should be
treated as undirected.}

\item{.bipartite}{If the layer specification is via a single
network with edge attributes and the network is unipartite, an
optional integer vector to specify which of the layers should be
treated as bipartite and how many \code{b1} vertices there are.}
}
\value{
A network object with layer metadata.
}
\description{
A function for specifying the LHS of a multilayer (a.k.a. multiplex
a.k.a. multirelational a.k.a. multivariate) ERGM.
}
\note{
If not all layers have the same bipartedness, all layers will
appear as unipartite to the statistics. However, an operator term
\code{\link[=ergm:ergm-terms]{S()}} can be used to construct a bipartite
subgraph of a unipartite graph.
}
\examples{

data(florentine)

# Method 1: list of networks
flo <- Layer(list(m = flomarriage, b = flobusiness))
ergm(flo ~ L(~edges, ~m)+L(~edges, ~b))

# Method 2: networks as arguments
flo <- Layer(m = flomarriage, b = flobusiness)
ergm(flo ~ L(~edges, ~m)+L(~edges, ~b))

# Method 3: edge attributes:
flo <- flomarriage | flobusiness
flo[,, names.eval="m"] <- as.matrix(flomarriage)
flo[,, names.eval="b"] <- as.matrix(flobusiness)
flo <- Layer(flo, c("m","b"))
ergm(flo ~ L(~edges, ~m)+L(~edges, ~b))

}
\seealso{
\link[=ergm-terms]{Help on model specification} for specific terms.
}
