% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_cond_sim.R
\name{marg_cond_sim}
\alias{marg_cond_sim}
\title{An experimental function for calculating gofN()-style Pearson residuals for arbitrary statistics.}
\usage{
marg_cond_sim(
  object,
  nsim = 1,
  obs.twostage = nsim/2,
  GOF = NULL,
  control = control.simulate.ergm(),
  ...
)
}
\description{
It should probably be moved to \code{ergm}, perhaps integrated into the \code{simulate} methods.
}
